Fixed page numbering
Ignore the original file name which is not important and used a simple
number with padded zeroes based on the total number of pages.

Additionally also wait for all threads to finish so that we can write
the images inside the zip file in the correct order (and not the order
of dictated by faster vs slower threads).